Refactor [#13] Todo 관련하여 리뷰사항 반영 및 패키지 변경작업 진행
