+++ morib/src/main/java/org/morib/server/domain/todo/application/FetchTodoServiceImpl.java
@@ -1,15 +1,13 @@
-package org.morib.server.api.timerView.service.fetch.todo;
+package org.morib.server.domain.todo.application;
 
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.todo.application.TodoGateway;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class FetchTodoServiceImpl implements FetchTodoService {
 
-    private final TodoGateway todoGateway;
 
     @Override
     public void fetch(LocalDate targetDate) {
