--- morib/src/main/java/org/morib/server/domain/todo/application/TodoGateway.java
@@ -1,17 +1,0 @@
-package org.morib.server.domain.todo.application;
-
-
-
-public interface TodoGateway {
-
-    void save();
-
-    void findById();
-
-    void findAll();
-
-    void deleteById();
-
-    void deleteAll();
-
-}
