+++ morib/src/main/java/org/morib/server/domain/todo/application/TodoManager.java
@@ -1,0 +1,5 @@
+package org.morib.server.domain.todo.application;
+
+public class TodoManager {
+
+}
